Align dtype handling with BF16 compute policy
Default model components and pipeline configs to BF16, while keeping FP32 only for scheduler/time/position math, tiled VAE blending, final waveform output, and the LTX-2.3 Vocoder+BWE precision island.

Let audio VAE decode and plain vocoder honor the configured compute dtype, preserve FP32 audio latent positions for the echo fix, and make debug output report the actual audio precision policy.

Document the current dtype policy, latent sidecar workflow, and mark the older audio issues note as historical.

+++ LTX_2_MLX/conditioning/tools.py
@@ -38,7 +38,7 @@ class VideoLatentTools:
 
     def create_initial_state(
         self,
-        dtype: mx.Dtype = mx.float32,
+        dtype: mx.Dtype = mx.bfloat16,
         initial_latent: Optional[mx.array] = None,
     ) -> LatentState:
         """
@@ -180,7 +180,7 @@ class AudioLatentTools:
 
     def create_initial_state(
         self,
-        dtype: mx.Dtype = mx.float32,
+        dtype: mx.Dtype = mx.bfloat16,
         initial_latent: Optional[mx.array] = None,
     ) -> LatentState:
         """
