+++ LTX_2_MLX/pipelines/a2vid_two_stage.py
@@ -64,7 +64,7 @@ class A2VidConfig:
     distilled_lora_config: Optional[LoRAConfig] = None
 
     tiling_config: Optional[TilingConfig] = None
-    dtype: mx.Dtype = mx.float16
+    dtype: mx.Dtype = mx.bfloat16
 
     # Audio params
     audio_vae_channels: int = 8
