+++ LTX_2_MLX/pipelines/ic_lora.py
@@ -258,7 +258,7 @@ class ICLoraConfig:
     tiling_config: Optional[TilingConfig] = None
 
     # Compute settings
-    dtype: mx.Dtype = mx.float32
+    dtype: mx.Dtype = mx.bfloat16
 
     def __post_init__(self):
         if self.num_frames % 8 != 1:
